requirements.txt
kivy==2.3.0
kivy-garden.mapview
requests
pytest
